Fix mobile viewport w/ Visualizer
The mobile viewport is now working as expected. The canvas was too large
which was causing problems for mobile. The animations reach 400px though
so on a super small phone it wont look right, but there is nothing I can
do about that. Additionally, that is not the intended use case for this
visualizer, so this becomes a non issue